apfs-label: get rid of mmap to support 16K pages
Apparently apfs-label does not work for 16K page sizes:

  https://github.com/linux-apfs/apfsprogs/issues/21

The problem is that the offsets need to be page-aligned. Just get rid of
mmap entirely, for a tool like this there is no performance cost in just
reading everything into a buffer.

The same problem exists in the mkfs and fsck though. I will have to deal
with it next.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>